Create `url-join` helper
https://github.com/DEFRA/water-abstraction-team/issues/26

Some of our services use an external dependency `url-join`, which can assemble urls using a series of regexes to manipulate the strings.

This package can cover all kinds of edge cases, however it is overkill for our purposes! So in the interests of reducing our external dependencies, we create our own drop-in replacement using node's native url and path handling.