Make experiment functions synchronous to fix tests
We updated the verison of `lab` when we moved it across from a dependency to a dev dependency. This new version has a breaking change where the test function of experiments cannot be async. Some of our tests broke as a result of this. We resolve this by simply removing `async` from the function definition.