Reimplement `deep-map` so we can remove dependencies
We use a package `deep-map` to iterate over an object/array and apply a function to each one. This also requires the dependency `es6-weak-map`. We don't need to handle all the different cases that `deep-map` can handle, which means we can write our own simplified implementation. We do this here and replace the use of the `deep-map` package with our own version. We also add unit testing for `digitise#transformNulls`, which makes use of `deepMap`.

Note that `digitise#transformNulls` is very similar to `nald#transformNull`; we could potentially rewrite the `digitise` method to call the `nald` method.