Reimplement `deep-map` dependency
We're pulling in a dependency `deep-map` along with a peer dependency `es6-weak-map`. `deep-map` is similar to `map` in that it iteratively applies a function, but it can also handle objects and recursion.

Looking at how we're using it, we can easily rewrite it ourselves, reducing our reliance on two dependencies (plus their sub-dependencies).