Reimplement `deep-map` dependency (#185)
* Reimplement `deep-map` dependency

We're pulling in a dependency `deep-map` along with a peer dependency `es6-weak-map`. `deep-map` is similar to `map` in that it iteratively applies a function, but it can also handle objects and recursion.

Looking at how we're using it, we can easily rewrite it ourselves, reducing our reliance on two dependencies (plus their sub-dependencies).

Our current unit tests handle some of our use of this. We fill the gaps by writing unit tests for `digitise#transformNulls`, which also makes use of it.

* Reimplement `deep-map` so we can remove dependencies

We use a package `deep-map` to iterate over an object/array and apply a function to each one. This also requires the dependency `es6-weak-map`. We don't need to handle all the different cases that `deep-map` can handle, which means we can write our own simplified implementation. We do this here and replace the use of the `deep-map` package with our own version. We also add unit testing for `digitise#transformNulls`, which makes use of `deepMap`.

Note that `digitise#transformNulls` is very similar to `nald#transformNull`; we could potentially rewrite the `digitise` method to call the `nald` method.

* Remove deep-map and es6-weak-map

Following our reimplemenation of it, we remove the `deep-map` dependency along with its dependency `es6-weak-map`