Remove `glob` dependency (#186)
We're pulling in the dependency `glob` but all we use it for is to read in an array of `*.json` filenames from a single folder, which is something we can easily do using native node functions. We therefore reimplement the `glob.sync` function and remove the dependency.

* Swap incorrect test files

We managed to end up with `test/deep-map` testing `digitise`, and `test/digitise` testing `deep-map`. We therefore swap the two files around to correct this.

* Reimplement `glob` and remove it

We are bringing in the `glob` package but only using it to read an array of *.json filenames in a single folder. We reimplement this functionality with native node functions and remove the `glob` dependency.

* Remove `glob` dependency

* Small formatting fix