Tidy up dependencies (#184)
* Tidy up dependencies

This PR aims to tidy up the dependencies we bring in. Some are being brought in as dependencies when they should be dev dependencies; some may not be needed at all; and in general we bring everything up to date where we can.

While doing this we noticed several dependencies which were being brought in for a single basic purpose (eg. `glob` supports nested folders and complex wildcarding but we were simply using it to find files in a single folder), or which we could have easily implemented ourselves. These have been resolved in other PRs:

- [Create `url-join` helper](https://github.com/DEFRA/water-abstraction-helpers/pull/183)
- [Reimplement `deep-map` dependency](https://github.com/DEFRA/water-abstraction-helpers/pull/185)
- [Remove `glob` dependency](https://github.com/DEFRA/water-abstraction-helpers/pull/186)
- [Reimplement `slugify` dependency](https://github.com/DEFRA/water-abstraction-helpers/pull/187)
- [Remove `pg-boss`](https://github.com/DEFRA/water-abstraction-helpers/pull/189)