Handle stacktrace only errors
https://eaflood.atlassian.net/browse/WATER-3845
https://github.com/DEFRA/water-abstraction-team/issues/45

We recently made a change to stop the madness of our logs getting jammed with the whole app codebase getting dumped everytime there was an error. All appeared good until an issue was raised when users attempted to cancel a bill run.

_Taken from [water-abstraction-service - src/modules/billing/services/batch-service.js](https://github.com/DEFRA/water-abstraction-service)_
```javascript
const deleteBatch = async (batch, internalCallingUser) => {
  if (!batch.canBeDeleted()) {
    throw new BatchStatusError(`Batch ${batch.id} cannot be deleted - status is ${batch.status}`)
  }

  try {
    await chargeModuleBillRunConnector.delete(batch.externalId)

    // These are populated at every stage in the bill run
    await newRepos.billingBatchChargeVersionYears.deleteByBatchId(batch.id, false)
    await newRepos.billingVolumes.deleteByBatchId(batch.id)

    // These tables are not yet populated at review stage in TPT
    await newRepos.billingTransactions.deleteByBatchId(batch.id)
    await newRepos.billingInvoiceLicences.deleteByBatchId(batch.id)
    await newRepos.billingInvoices.deleteByBatchId(batch.id, false)
    await newRepos.billingBatches.delete(batch.id)

    await saveEvent('billing-batch:cancel', 'delete', internalCallingUser, batch)
  } catch (err) {
    logger.error('Failed to delete the batch', err.stack, batch)
    await saveEvent('billing-batch:cancel', 'error', internalCallingUser, batch)
    await setStatus(batch.id, BATCH_STATUS.error)
    throw err
  }
}
```

From looking at the logs it seems our request to the charging module API to delete the bill run is not being handled gracefully. If there is any authentication issues, for example the cognito token has expired, it gets into a mess. The bill run does get deleted in the CHA, but we also seem to go into a cycle of retrying. This will cause the CHA API to return a `4xx` which then causes an exception to be thrown on the WRLS side.

From looking at the data this isn't the first time it's happened. But previously the batch would get flagged as errored, after the failure had been logged (who cares about deleting the data!😜). In this case the logger code in this repo is expecting an error object and is throwing an error when it doesn't get one. This goes all the way to the top, causing the event and status updates to never get called in the above code.

We're not going back (we only tracked this down _because_ the logs are now approaching being usable). But we need to make a change to cater for this scenario.