Add tests for calls we haven't covered
It might _look_ like we have logging an error covered. But as our example shows, we're not calling `logger.log('error', 'scary message)`, we're calling `logger.error('scary message')`.

So, have added some tests to cover how the logger is actually used 😩🤦. And suprise-surprise, we now have a failing test!