Update packages
A lot of the legacy projects use this repo. There have already been a lot of dependency updates made to this repo since it's last release (Nov 2022).

In this PR we are going to update any dependencies that still are out of date. Then it looks like if we tag it and then create a release in GitHub it will automatically push a new package to npm.

If it publishes successfully we will then be updating all of the legacy projects that have this helper as a dependency to it's latest version.